requirements.txt
telethon==1.24.0
tqdm==4.62.3
python-dotenv